Deleted last renames.
